requirements.txt
asgiref==3.5.2
channels==3.0.5
channels-redis==3.4.1
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.21.3
graphene==2.1.9
graphene-django==2.15.0
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.2.1
redis==4.3.4
